refactor: encapsulate anime state in component and improve upsert logic
- Extracted and encapsulated anime state management in the WatchStatusSelector component.
- Refactored AnimeDescriptionPage to use the new component.
- Created UpdateAnimeStatus function to handle anime status updates.
- Created RemoveAnimeStatus function to handle removal of anime status.
- Created ProfilePageClient layout to manage user profile data.
- Added onConflict parameter in upsert to avoid duplicates and correctly handle the unique constraint in Supabase.
- Improved code architecture and maintainability.
